+++ package.json
@@ -1,14 +1,14 @@
 {
   "name": "chaotic-app",
   "type": "module",
-  "version": "7.0.0",
+  "version": "8.0.0",
   "private": true,
-  "packageManager": "pnpm@10.20.0",
+  "packageManager": "pnpm@11.1.2",
   "engines": {
     "node": "^24.x"
   },
   "scripts": {
-    "build": "papi --whitelist .papi/whitelist.ts && nuxt build",
+    "build": "papi && nuxt build",
     "dev": "nuxt dev",
     "generate": "nuxt generate",
     "generate:providers": "node scripts/generate-providers.mjs",
@@ -19,69 +19,66 @@
     "prepare": "husky",
     "test": "vitest run",
     "test:watch": "vitest",
-    "typecheck": "papi --whitelist .papi/whitelist.ts && nuxi typecheck"
+    "typecheck": "papi && nuxi typecheck"
   },
   "dependencies": {
-    "@apollo/client": "^3.13.8",
-    "@fal-ai/client": "^1.6.2",
-    "@iconify-json/lucide": "^1.2.57",
-    "@iconify-json/simple-icons": "^1.2.43",
+    "@apollo/client": "^4.1.9",
+    "@fal-ai/client": "^1.10.1",
+    "@iconify-json/lucide": "^1.2.108",
+    "@iconify-json/simple-icons": "^1.2.82",
     "@kodadot1/hyperdata": "0.0.2-rc.1",
     "@kodadot1/static": "0.0.8-rc.0",
-    "@nuxt/fonts": "0.11.4",
-    "@nuxt/icon": "2.1.0",
-    "@nuxt/ui": "^4.1.0",
-    "@nuxtjs/i18n": "^10.1.2",
-    "@nuxtjs/seo": "^3.4.0",
+    "@nuxt/fonts": "0.14.0",
+    "@nuxt/icon": "2.2.2",
+    "@nuxt/ui": "^4.7.1",
+    "@nuxtjs/i18n": "^10.3.0",
+    "@nuxtjs/seo": "^5.1.3",
     "@paraport/vue": "0.2.1",
-    "@pinia/nuxt": "0.11.1",
-    "@polkadot/util-crypto": "^13.5.7",
-    "@posthog/nuxt": "^1.5.84",
-    "@tanstack/vue-query": "^5.83.0",
-    "@unhead/vue": "^2.0.11",
-    "@vueuse/core": "^13.9.0",
-    "@wagmi/vue": "^0.4.15",
+    "@pinia/nuxt": "0.11.3",
+    "@polkadot/util-crypto": "^14.0.3",
+    "@posthog/nuxt": "^1.7.31",
+    "@takumi-rs/core": "^1.1.2",
+    "@tanstack/vue-query": "^5.100.10",
+    "@unhead/vue": "^3.1.0",
+    "@vueuse/core": "^14.3.0",
+    "@wagmi/vue": "^0.5.16",
     "chart.js": "^4.5.1",
     "client-zip": "^2.5.0",
     "date-fns": "^4.1.0",
-    "dedot": "^0.14.1",
-    "gql.tada": "^1.8.11",
+    "dedot": "^1.3.0",
+    "gql.tada": "^1.9.2",
     "jdenticon": "^3.3.0",
-    "markdown-it": "^14.1.0",
-    "nuxt": "^4.2.0",
-    "nuxt-svgo": "4.2.6",
-    "ofetch": "1.4.1",
-    "pinia": "^3.0.3",
-    "pinia-plugin-persistedstate": "^4.4.1",
-    "polkadot-api": "^1.20.0",
+    "markdown-it": "^14.1.1",
+    "nuxt": "^4.4.5",
+    "nuxt-svgo": "5.3.0",
+    "ofetch": "1.5.1",
+    "pinia": "^3.0.4",
+    "pinia-plugin-persistedstate": "^4.7.1",
+    "polkadot-api": "^2.1.3",
     "prismjs": "^1.30.0",
     "stats.js": "^0.17.0",
-    "try": "1.0.0-beta.10",
-    "unzipit": "^1.4.3",
-    "valibot": "^1.1.0",
-    "vue": "^3.5.22",
+    "try": "1.0.3",
+    "unzipit": "^2.0.1",
+    "valibot": "^1.4.0",
+    "viem": "~2.49.3",
+    "vue": "^3.5.34",
     "vue-chartjs": "^5.3.3",
     "vue-dompurify-html": "^5.3.0"
   },
   "devDependencies": {
-    "@antfu/eslint-config": "^6.1.0",
-    "@gql.tada/vue-support": "^1.0.1",
-    "@nuxt/eslint": "^1.14.0",
-    "@polkadot/apps-config": "^0.169.1",
-    "@types/lodash": "^4.17.20",
+    "@antfu/eslint-config": "^9.0.0",
+    "@gql.tada/vue-support": "^1.0.2",
+    "@nuxt/eslint": "^1.15.2",
+    "@polkadot/apps-config": "^0.171.1",
+    "@types/lodash": "^4.17.24",
     "@types/markdown-it": "^14.1.2",
-    "@types/prismjs": "^1.26.5",
-    "baseline-browser-mapping": "^2.9.19",
-    "eslint": "^9.38.0",
+    "@types/prismjs": "^1.26.6",
+    "baseline-browser-mapping": "^2.10.30",
+    "eslint": "^10.4.0",
     "husky": "^9.1.7",
-    "pino-pretty": "^13.1.1",
-    "typescript": "^5.8.3",
-    "vitest": "^4.0.18",
-    "vue-tsc": "^3.2.3"
-  },
-  "pnpm": {
-    "overrides": {
-      "@noble/hashes": "1.8.0"
-    }
+    "pino-pretty": "^13.1.3",
+    "typescript": "^5.9.3",
+    "vitest": "^4.1.6",
+    "vue-tsc": "^3.2.9"
   }
 }
