+++ pnpm-workspace.yaml
@@ -1,0 +1,16 @@
+allowBuilds:
+  '@parcel/watcher': true
+  '@posthog/cli': true
+  bufferutil: true
+  core-js: true
+  es5-ext: true
+  esbuild: true
+  protobufjs: true
+  sharp: true
+  unrs-resolver: true
+  utf-8-validate: true
+  vue-demi: true
+shellEmulator: true
+trustPolicy: no-downgrade
+overrides:
+  '@noble/hashes': 1.8.0
